代码更新

+++ discuz/cleancache.php
@@ -5,7 +5,7 @@ $path = dirname(__FILE__) . DIRECTORY_SEPARATOR . "data";
 $uc_path = dirname(__FILE__) . DIRECTORY_SEPARATOR . "uc_server" . DIRECTORY_SEPARATOR . "data";
 
 // 清理discuz缓存文件
-function cleanCacheFile($dir)
+function cleanCacheFile($dir, $type = 1)
 {
     $handle = opendir($dir);
     while ($file = readdir($handle)) {
@@ -13,11 +13,23 @@ function cleanCacheFile($dir)
             $f = $dir . DIRECTORY_SEPARATOR . $file;
             if (is_file($f)) {
 //                echo "|--" . $file . "<br>";
-                unlink($f);
+                if ($type === 1) {
+                    unlink($f);
+                }
             } else {
+                if ($type === 2) {
+                    echo "目录名：" . $file . "<br>";
+                    $myfile = $f . DIRECTORY_SEPARATOR . "index.htm";
+                    $fileRes = fopen($myfile, "w+");
+                }
 //                echo "--" . $file . "<br>";
                 if ($file !== "attachment") {
-                    cleanCacheFile($f);
+                    if ($type === 1) {
+                        cleanCacheFile($f);
+                    }
+                    if ($type === 2) {
+                        cleanCacheFile($f, 2);
+                    }
                 }
             }
         }
@@ -28,3 +40,8 @@ cleanCacheFile($path);
 echo "清理完成data目录缓存。<br>";
 //cleanCacheFile($uc_path);
 //echo "清理完成uc_server/data目录缓存。<br>";
+
+
+// 往data目录，每个子目录创建一个index.htm空文件
+cleanCacheFile($path, 2);
+echo "完成在每个目录下创建空文件index.htm。";
